test: add comprehensive test coverage from test-coverage-audit
- Add test_core.py - Core data structure tests
- Add test_logistic_coverage.py - Logistic regression coverage tests
- Add test_engine_iterators.py - Engine chunk iteration tests
- Add test_plink_iterators.py - PLINK I/O iterator tests
- Add test_tabular.py - Tabular data loading tests

+++ tests/test_engine_iterators.py
@@ -112,7 +112,7 @@ def test_iter_logistic_output_frames_passes_no_missing_constants() -> None:
             iteration_count=jnp.array([3], dtype=jnp.int32),
         ),
         allele_one_frequency=jnp.array([0.25]),
-        observation_count=jnp.array([2], dtype=jnp.int32),
+        observation_count=jnp.array([2], dtype=jnp.int64),
     )
     no_missing_constants = object()
